Kernel: Remove the tty_writeready() function
I added tty_writeready() to socz80-UZI as a way to get better
concurrency by causing a process writing to a busy UART port to sleep,
thus allowing a second process to run, and later waking up the sleeping
process once a UART interrupt arrives to indicate it is ready for
another byte.

This worked well enough on socz80-UZI with its 128MHz processor, but on
real Z80 hardware this scheme actually hurts performance at most
reasonable baud rates; the necessary overhead of sleeping the process,
switching to another process, taking an interrupt and switching back
takes much longer than just waiting for the UART to finish transmitting.

This patch removes tty_writeready in the core kernel and for each
platform. Note that an individual platform which wishes to retain the
old "sleep on busy UART" behaviour can still do so simply by
implementing this inside the platform tty_putc() routine.